requirements.txt
streamlit
pytesseract
Pillow
pandas
python-dotenv 